Use the same recipe for both versions
