+++ recipes/msgpack23/all/conandata.yml
@@ -1,4 +1,7 @@
 sources:
+  "3.0":
+    url: "https://github.com/rwindegger/msgpack23/archive/refs/tags/v3.0.tar.gz"
+    sha256: "706CB1A623B9FB178F1A14C00322B147165B57E557DF9777A5825FF32052EB5F"
   "2.1":
     url: "https://github.com/rwindegger/msgpack23/archive/refs/tags/v2.1.tar.gz"
     sha256: "9CE1E294518AA76CAC50F778A359AED17A0DAA0D8DC4C1F94CD4F12438B3606C"
