+++ recipes/msgpack23/all/test_package/conanfile.py
@@ -1,18 +1,24 @@
 from conan import ConanFile
 from conan.tools.build import can_run
-from conan.tools.cmake import cmake_layout, CMake
+from conan.tools.cmake import cmake_layout, CMake, CMakeToolchain
 import os
 
 
 class TestPackageConan(ConanFile):
     settings = "os", "arch", "compiler", "build_type"
-    generators = "CMakeDeps", "CMakeToolchain"
+    generators = "CMakeDeps"
 
     def layout(self):
         cmake_layout(self)
 
     def requirements(self):
         self.requires(self.tested_reference_str)
+
+    def generate(self):
+        tc = CMakeToolchain(self)
+        print(type(self.dependencies[self.tested_reference_str].ref.version))
+        tc.cache_variables["IS_NEW_API"] = self.dependencies[self.tested_reference_str].ref.version >= "3.0"
+        tc.generate()
 
     def build(self):
         cmake = CMake(self)
