+++ recipes/msgpack23/all/test_package/test_package_old.cpp
@@ -1,13 +1,10 @@
 #include <msgpack23/msgpack23.h>
-#include <vector>
-#include <iterator>
 
 int main(void) {
-	std::vector<std::byte> data{};
-	auto inserter = std::back_inserter(data);
-    msgpack23::Packer packer {inserter};
+	
+    msgpack23::Packer packer {};
 	std::uint8_t const expected = 42;
-	packer(expected);
+	auto data = packer(expected);
 	msgpack23::Unpacker unpacker {data};
 	std::int8_t actual {};
 	unpacker(actual);
