+++ recipes/msgpack23/config.yml
@@ -1,5 +1,5 @@
 versions:
+  "3.0":
+    folder: all
   "2.1":
     folder: all
-  "3.0":
-    folder: v3.0
