--- recipes/msgpack23/v3.0/test_package/conanfile.py
@@ -1,25 +1,0 @@
-from conan import ConanFile
-from conan.tools.build import can_run
-from conan.tools.cmake import cmake_layout, CMake
-import os
-
-
-class TestPackageConan(ConanFile):
-    settings = "os", "arch", "compiler", "build_type"
-    generators = "CMakeDeps", "CMakeToolchain"
-
-    def layout(self):
-        cmake_layout(self)
-
-    def requirements(self):
-        self.requires(self.tested_reference_str)
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
-
-    def test(self):
-        if can_run(self):
-            bin_path = os.path.join(self.cpp.build.bindir, "test_package")
-            self.run(bin_path, env="conanrun")
